+++ src/pages/MainPage.tsx
@@ -1,76 +1,18 @@
-import { DragDropContext, DropResult } from "react-beautiful-dnd";
+import styled from "styled-components";
+
 import LeftSidePage from "./LeftSidePage";
 import RightSidePage from "./RightSidePage";
 import ToDoPage from "./ToDoPage";
-import styled from "styled-components";
-import { useRecoilValue, useSetRecoilState } from "recoil";
-import {
-  isDraggingState,
-  isOverTrashCanState,
-  toDosByDateSelector,
-  toDosState,
-} from "../models/atoms";
-import { findSameId } from "../utils/RecoilFunctions";
 
 type pageProps = { className: string };
 
 function MainPage() {
-  const setTodos = useSetRecoilState(toDosState);
-  const toDosByDate = useRecoilValue(toDosByDateSelector);
-  const setIsDragging = useSetRecoilState(isDraggingState);
-  const setIsOverTrashCan = useSetRecoilState(isOverTrashCanState);
-  const dragEndHandler = ({ destination, draggableId }: DropResult) => {
-    setIsDragging(false);
-    setIsOverTrashCan(false);
-    if (!destination) {
-      return;
-    }
-    if (destination.droppableId === "todos") {
-      setTodos((allTodos) => {
-        const copyTodos = [...allTodos];
-
-        const sourceIndex = allTodos.findIndex((element) =>
-          findSameId(element, +draggableId)
-        );
-        const moveTodo = copyTodos[sourceIndex];
-        const destinationIndex = allTodos.findIndex((element) =>
-          findSameId(element, toDosByDate[destination.index].id)
-        );
-        copyTodos.splice(sourceIndex, 1);
-        copyTodos.splice(destinationIndex, 0, moveTodo);
-
-        return copyTodos;
-      });
-    } else if (destination.droppableId === "trashcan") {
-      setTodos((allTodos) => {
-        const copyTodos = [...allTodos];
-
-        const sourceIndex = allTodos.findIndex((element) =>
-          findSameId(element, +draggableId)
-        );
-        copyTodos.splice(sourceIndex, 1);
-        return copyTodos;
-      });
-    }
-  };
   return (
-    <DragDropContext
-      onDragEnd={dragEndHandler}
-      onBeforeCapture={() => setIsDragging(true)}
-      onDragUpdate={(update) => {
-        if (update.destination?.droppableId === "trashcan") {
-          setIsOverTrashCan(true);
-        } else {
-          setIsOverTrashCan(false);
-        }
-      }}
-    >
-      <Wrapper>
-        <LeftSidePageWrapper className="" />
-        <ToDoPageWrapper className="" />
-        <RightSidePageWrapper className="" />
-      </Wrapper>
-    </DragDropContext>
+    <Wrapper>
+      <LeftSidePageWrapper className="" />
+      <ToDoPageWrapper className="" />
+      <RightSidePageWrapper className="" />
+    </Wrapper>
   );
 }
